Migrate src/index.js
