+++ apps/backend/src/config/db.ts
@@ -1,12 +1,12 @@
 import mongoose from 'mongoose';
 
 export const connectDB = async () => {
-  console.log('🔍 Debug - Environment variables:');
+  console.log('Debug - Environment variables:');
   console.log('MONGO_URI:', process.env.MONGO_URI);
   console.log('PORT:', process.env.PORT);
 
   if (!process.env.MONGO_URI) {
-    console.error('❌ MONGO_URI is not defined in environment variables');
+    console.error('MONGO_URI is not defined in environment variables');
     console.error(
       '💡 Make sure your .env file is in the correct location and contains MONGO_URI',
     );
@@ -14,11 +14,11 @@ export const connectDB = async () => {
   }
 
   try {
-    console.log(`🔄 Connecting to MongoDB...`);
+    console.log(`Connecting to MongoDB...`);
     await mongoose.connect(process.env.MONGO_URI as string);
-    console.log('✅ MongoDB connected successfully');
+    console.log('MongoDB connected successfully');
   } catch (err) {
-    console.error('❌ MongoDB connection error:', err);
+    console.error('MongoDB connection error:', err);
     process.exit(1);
   }
 };
